+++ CHANGES.md
@@ -10,6 +10,9 @@
 * Removed previously deprecated class: `org.jsoup.UncheckedIOException` (replace with `java.io.UncheckedIOException`);
   method `Element Element#forEach(Consumer)` to
   `void Element#forEach(Consumer())`. [2246](https://github.com/jhy/jsoup/pull/2246)
+* Deprecated the methods `Document#updateMetaCharsetElement(bool)` and `#Document#updateMetaCharsetElement()`, as the
+  setting had no effect. When `Document#charset(Charset)` is called, the document's meta charset or XML encoding
+  instruction is always set. [2247](https://github.com/jhy/jsoup/pull/2247)
 
 ### Improvements